+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - On macOS, `VirtualKeyCode::RWin` and `VirtualKeyCode::LWin` are no longer switched.
 - On macOS, windows without decorations can once again be resized.
 - Fixed race conditions when creating an `EventsLoop` on X11, most commonly manifesting as "[xcb] Unknown sequence number while processing queue".
+- On macOS, `MouseInput` events are only generated if the click occurs within the window's client area.
 
 # Version 0.15.0 (2018-05-22)